split network implementations into separate files
