+++ include/sorting_network_cpp/networks/size_optimized_sort.h
@@ -1,0 +1,370 @@
+// This file is part of the sorting_network_cpp (https://github.com/quxflux/sorting_network_cpp).
+// Copyright (c) 2022 Lukas Riebel.
+//
+// sorting_network_cpp is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// sorting_network_cpp is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+#pragma once
+
+#include <sorting_network_cpp/networks/common.h>
+
+namespace quxflux::sorting_net::detail
+{
+  namespace sorter_hunter
+  {
+    template<std::size_t A, std::size_t B>
+    using P = sorting_net::cas_node<A, B>;
+
+    template<typename... Pairs>
+    using L = sorting_net::layer<Pairs...>;
+
+    template<typename... Layers>
+    using N = sorting_net::net<Layers...>;
+
+    // clang-format off
+    using Networks = std::tuple<
+                /* 1 */  N<L<>>,
+                /* 2 */  N<L<P<0,1>>>,
+                /* 3 */  N<L<P<0,2>>,
+                          L<P<0,1>>,
+                          L<P<1,2>>>,
+                /* 4 */  N<L<P<0,2>,P<1,3>>,
+                          L<P<0,1>,P<2,3>>,
+                          L<P<1,2>>>,
+                /* 5 */  N<L<P<0,3>,P<1,4>>,
+                          L<P<0,2>,P<1,3>>,
+                          L<P<0,1>,P<2,4>>,
+                          L<P<1,2>,P<3,4>>,
+                          L<P<2,3>>>,
+                /* 6 */  N<L<P<0,5>,P<1,3>,P<2,4>>,
+                          L<P<1,2>,P<3,4>>,
+                          L<P<0,3>,P<2,5>>,
+                          L<P<0,1>,P<2,3>,P<4,5>>,
+                          L<P<1,2>,P<3,4>>>,
+                /* 7 */  N<L<P<0,6>,P<2,3>,P<4,5>>,
+                          L<P<0,2>,P<1,4>,P<3,6>>,
+                          L<P<0,1>,P<2,5>,P<3,4>>,
+                          L<P<1,2>,P<4,6>>,
+                          L<P<2,3>,P<4,5>>,
+                          L<P<1,2>,P<3,4>,P<5,6>>>,
+                /* 8 */  N<L<P<0,2>,P<1,3>,P<4,6>,P<5,7>>,
+                          L<P<0,4>,P<1,5>,P<2,6>,P<3,7>>,
+                          L<P<0,1>,P<2,3>,P<4,5>,P<6,7>>,
+                          L<P<2,4>,P<3,5>>,
+                          L<P<1,4>,P<3,6>>,
+                          L<P<1,2>,P<3,4>,P<5,6>>>,
+                /* 9 */  N<L<P<0,3>,P<1,7>,P<2,5>,P<4,8>>,
+                          L<P<0,7>,P<2,4>,P<3,8>,P<5,6>>,
+                          L<P<0,2>,P<1,3>,P<4,5>,P<7,8>>,
+                          L<P<1,4>,P<3,6>,P<5,7>>,
+                          L<P<0,1>,P<2,4>,P<3,5>,P<6,8>>,
+                          L<P<2,3>,P<4,5>,P<6,7>>,
+                          L<P<1,2>,P<3,4>,P<5,6>>>,
+                /* 10 */ N<L<P<0,8>,P<1,9>,P<2,7>,P<3,5>,P<4,6>>,
+                          L<P<0,2>,P<1,4>,P<5,8>,P<7,9>>,
+                          L<P<0,3>,P<2,4>,P<5,7>,P<6,9>>,
+                          L<P<0,1>,P<3,6>,P<8,9>>,
+                          L<P<1,5>,P<2,3>,P<4,8>,P<6,7>>,
+                          L<P<1,2>,P<3,5>,P<4,6>,P<7,8>>,
+                          L<P<2,3>,P<4,5>,P<6,7>>,
+                          L<P<3,4>,P<5,6>>>,
+                /* 11 */ N<L<P<0,9>,P<1,6>,P<2,4>,P<3,7>,P<5,8>>,
+                          L<P<0,1>,P<3,5>,P<4,10>,P<6,9>,P<7,8>>,
+                          L<P<1,3>,P<2,5>,P<4,7>,P<8,10>>,
+                          L<P<0,4>,P<1,2>,P<3,7>,P<5,9>,P<6,8>>,
+                          L<P<0,1>,P<2,6>,P<4,5>,P<7,8>,P<9,10>>,
+                          L<P<2,4>,P<3,6>,P<5,7>,P<8,9>>,
+                          L<P<1,2>,P<3,4>,P<5,6>,P<7,8>>,
+                          L<P<2,3>,P<4,5>,P<6,7>>>,
+                /* 12 */ N<L<P<0,8>,P<1,7>,P<2,6>,P<3,11>,P<4,10>,P<5,9>>,
+                          L<P<0,1>,P<2,5>,P<3,4>,P<6,9>,P<7,8>,P<10,11>>,
+                          L<P<0,2>,P<1,6>,P<5,10>,P<9,11>>,
+                          L<P<0,3>,P<1,2>,P<4,6>,P<5,7>,P<8,11>,P<9,10>>,
+                          L<P<1,4>,P<3,5>,P<6,8>,P<7,10>>,
+                          L<P<1,3>,P<2,5>,P<6,9>,P<8,10>>,
+                          L<P<2,3>,P<4,5>,P<6,7>,P<8,9>>,
+                          L<P<4,6>,P<5,7>>,
+                          L<P<3,4>,P<5,6>,P<7,8>>>,
+                /* 13 */ N<L<P<0,12>,P<1,10>,P<2,9>,P<3,7>,P<5,11>,P<6,8>>,
+                          L<P<1,6>,P<2,3>,P<4,11>,P<7,9>,P<8,10>>,
+                          L<P<0,4>,P<1,2>,P<3,6>,P<7,8>,P<9,10>,P<11,12>>,
+                          L<P<4,6>,P<5,9>,P<8,11>,P<10,12>>,
+                          L<P<0,5>,P<3,8>,P<4,7>,P<6,11>,P<9,10>>,
+                          L<P<0,1>,P<2,5>,P<6,9>,P<7,8>,P<10,11>>,
+                          L<P<1,3>,P<2,4>,P<5,6>,P<9,10>>,
+                          L<P<1,2>,P<3,4>,P<5,7>,P<6,8>>,
+                          L<P<2,3>,P<4,5>,P<6,7>,P<8,9>>,
+                          L<P<3,4>,P<5,6>>>,
+                /* 14 */ N<L<P<0,6>,P<1,11>,P<2,12>,P<3,10>,P<4,5>,P<7,13>,P<8,9>>,
+                          L<P<1,2>,P<3,7>,P<4,8>,P<5,9>,P<6,10>,P<11,12>>,
+                          L<P<0,4>,P<1,3>,P<5,6>,P<7,8>,P<9,13>,P<10,12>>,
+                          L<P<0,1>,P<2,9>,P<3,7>,P<4,11>,P<6,10>,P<12,13>>,
+                          L<P<2,5>,P<4,7>,P<6,9>,P<8,11>>,
+                          L<P<1,2>,P<3,4>,P<6,7>,P<9,10>,P<11,12>>,
+                          L<P<1,3>,P<2,4>,P<5,6>,P<7,8>,P<9,11>,P<10,12>>,
+                          L<P<2,3>,P<4,7>,P<6,9>,P<10,11>>,
+                          L<P<4,5>,P<6,7>,P<8,9>>,
+                          L<P<3,4>,P<5,6>,P<7,8>,P<9,10>>>,
+                /* 15 */ N<L<P<1,2>,P<3,10>,P<4,14>,P<5,8>,P<6,13>,P<7,12>,P<9,11>>,
+                          L<P<0,14>,P<1,5>,P<2,8>,P<3,7>,P<6,9>,P<10,12>,P<11,13>>,
+                          L<P<0,7>,P<1,6>,P<2,9>,P<4,10>,P<5,11>,P<8,13>,P<12,14>>,
+                          L<P<0,6>,P<2,4>,P<3,5>,P<7,11>,P<8,10>,P<9,12>,P<13,14>>,
+                          L<P<0,3>,P<1,2>,P<4,7>,P<5,9>,P<6,8>,P<10,11>,P<12,13>>,
+                          L<P<0,1>,P<2,3>,P<4,6>,P<7,9>,P<10,12>,P<11,13>>,
+                          L<P<1,2>,P<3,5>,P<8,10>,P<11,12>>,
+                          L<P<3,4>,P<5,6>,P<7,8>,P<9,10>>,
+                          L<P<2,3>,P<4,5>,P<6,7>,P<8,9>,P<10,11>>,
+                          L<P<5,6>,P<7,8>>>,
+                /* 16 */ N<L<P<0,13>,P<1,12>,P<2,15>,P<3,14>,P<4,8>,P<5,6>,P<7,11>,P<9,10>>,
+                          L<P<0,5>,P<1,7>,P<2,9>,P<3,4>,P<6,13>,P<8,14>,P<10,15>,P<11,12>>,
+                          L<P<0,1>,P<2,3>,P<4,5>,P<6,8>,P<7,9>,P<10,11>,P<12,13>,P<14,15>>,
+                          L<P<0,2>,P<1,3>,P<4,10>,P<5,11>,P<6,7>,P<8,9>,P<12,14>,P<13,15>>,
+                          L<P<1,2>,P<3,12>,P<4,6>,P<5,7>,P<8,10>,P<9,11>,P<13,14>>,
+                          L<P<1,4>,P<2,6>,P<5,8>,P<7,10>,P<9,13>,P<11,14>>,
+                          L<P<2,4>,P<3,6>,P<9,12>,P<11,13>>,
+                          L<P<3,5>,P<6,8>,P<7,9>,P<10,12>>,
+                          L<P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,12>>,
+                          L<P<6,7>,P<8,9>>>,
+                /* 17 */ N<L<P<0,11>,P<1,15>,P<2,10>,P<3,5>,P<4,6>,P<8,12>,P<9,16>,P<13,14>>,
+                          L<P<0,6>,P<1,13>,P<2,8>,P<4,14>,P<5,15>,P<7,11>>,
+                          L<P<0,8>,P<3,7>,P<4,9>,P<6,16>,P<10,11>,P<12,14>>,
+                          L<P<0,2>,P<1,4>,P<5,6>,P<7,13>,P<8,9>,P<10,12>,P<11,14>,P<15,16>>,
+                          L<P<0,3>,P<2,5>,P<6,11>,P<7,10>,P<9,13>,P<12,15>,P<14,16>>,
+                          L<P<0,1>,P<3,4>,P<5,10>,P<6,9>,P<7,8>,P<11,15>,P<13,14>>,
+                          L<P<1,2>,P<3,7>,P<4,8>,P<6,12>,P<11,13>,P<14,15>>,
+                          L<P<1,3>,P<2,7>,P<4,5>,P<9,11>,P<10,12>,P<13,14>>,
+                          L<P<2,3>,P<4,6>,P<5,7>,P<8,10>>,
+                          L<P<3,4>,P<6,8>,P<7,9>,P<10,12>>,
+                          L<P<5,6>,P<7,8>,P<9,10>,P<11,12>>,
+                          L<P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>>>,
+                /* 18 */ N<L<P<0,1>,P<2,3>,P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<14,15>,P<16,17>>,
+                          L<P<1,5>,P<2,6>,P<3,7>,P<4,10>,P<8,16>,P<9,17>,P<12,14>,P<13,15>>,
+                          L<P<0,8>,P<1,10>,P<2,12>,P<3,14>,P<6,13>,P<7,15>,P<9,16>,P<11,17>>,
+                          L<P<0,4>,P<1,9>,P<5,17>,P<8,11>,P<10,16>>,
+                          L<P<0,2>,P<1,6>,P<4,10>,P<5,9>,P<14,16>,P<15,17>>,
+                          L<P<1,2>,P<3,10>,P<4,12>,P<5,7>,P<6,14>,P<9,13>,P<15,16>>,
+                          L<P<3,8>,P<5,12>,P<7,11>,P<9,10>>,
+                          L<P<3,4>,P<6,8>,P<7,14>,P<9,12>,P<11,13>>,
+                          L<P<1,3>,P<2,4>,P<7,9>,P<8,12>,P<11,15>,P<13,16>>,
+                          L<P<2,3>,P<4,5>,P<6,7>,P<10,11>,P<12,14>,P<13,15>>,
+                          L<P<4,6>,P<5,8>,P<9,10>,P<11,14>>,
+                          L<P<3,4>,P<5,7>,P<8,9>,P<10,12>,P<13,14>>,
+                          L<P<5,6>,P<7,8>,P<9,10>,P<11,12>>>,
+                /* 19 */ N<L<P<0,12>,P<1,4>,P<2,8>,P<3,5>,P<6,17>,P<7,11>,P<9,14>,P<10,13>,P<15,16>>,
+                          L<P<0,2>,P<1,7>,P<3,6>,P<4,11>,P<5,17>,P<8,12>,P<10,15>,P<13,16>,P<14,18>>,
+                          L<P<3,10>,P<4,14>,P<5,15>,P<6,13>,P<7,9>,P<11,17>,P<16,18>>,
+                          L<P<0,7>,P<1,10>,P<4,6>,P<9,15>,P<11,16>,P<12,17>,P<13,14>>,
+                          L<P<0,3>,P<2,6>,P<5,7>,P<8,11>,P<12,16>>,
+                          L<P<1,8>,P<2,9>,P<3,4>,P<6,15>,P<7,13>,P<10,11>,P<12,18>>,
+                          L<P<1,3>,P<2,5>,P<6,9>,P<7,12>,P<8,10>,P<11,14>,P<17,18>>,
+                          L<P<0,1>,P<2,3>,P<4,8>,P<6,10>,P<9,12>,P<14,15>,P<16,17>>,
+                          L<P<1,2>,P<5,8>,P<6,7>,P<9,11>,P<10,13>,P<14,16>,P<15,17>>,
+                          L<P<3,6>,P<4,5>,P<7,9>,P<8,10>,P<11,12>,P<13,14>,P<15,16>>,
+                          L<P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,13>,P<12,14>>,
+                          L<P<2,3>,P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<14,15>>>,
+                /* 20 */ N<L<P<0,3>,P<1,7>,P<2,5>,P<4,8>,P<6,9>,P<10,13>,P<11,15>,P<12,18>,P<14,17>,P<16,19>>,
+                          L<P<0,14>,P<1,11>,P<2,16>,P<3,17>,P<4,12>,P<5,19>,P<6,10>,P<7,15>,P<8,18>,P<9,13>>,
+                          L<P<0,4>,P<1,2>,P<3,8>,P<5,7>,P<11,16>,P<12,14>,P<15,19>,P<17,18>>,
+                          L<P<1,6>,P<2,12>,P<3,5>,P<4,11>,P<7,17>,P<8,15>,P<13,18>,P<14,16>>,
+                          L<P<0,1>,P<2,6>,P<7,10>,P<9,12>,P<13,17>,P<18,19>>,
+                          L<P<1,6>,P<5,9>,P<7,11>,P<8,12>,P<10,14>,P<13,18>>,
+                          L<P<3,5>,P<4,7>,P<8,10>,P<9,11>,P<12,15>,P<14,16>>,
+                          L<P<1,3>,P<2,4>,P<5,7>,P<6,10>,P<9,13>,P<12,14>,P<15,17>,P<16,18>>,
+                          L<P<1,2>,P<3,4>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<15,16>,P<17,18>>,
+                          L<P<2,3>,P<4,6>,P<5,8>,P<7,9>,P<10,12>,P<11,14>,P<13,15>,P<16,17>>,
+                          L<P<4,5>,P<6,8>,P<7,10>,P<9,12>,P<11,13>,P<14,15>>,
+                          L<P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>>>,
+                /* 21 */ N<L<P<0,7>,P<1,10>,P<3,5>,P<4,8>,P<6,13>,P<9,19>,P<11,14>,P<12,17>,P<15,16>,P<18,20>>,
+                          L<P<0,11>,P<1,15>,P<2,12>,P<3,4>,P<5,8>,P<6,9>,P<7,14>,P<10,16>,P<13,19>,P<17,20>>,
+                          L<P<0,6>,P<1,3>,P<2,18>,P<4,15>,P<5,10>,P<8,16>,P<11,17>,P<12,13>,P<14,20>>,
+                          L<P<2,6>,P<5,12>,P<7,18>,P<8,14>,P<9,11>,P<10,17>,P<13,19>,P<16,20>>,
+                          L<P<1,2>,P<4,7>,P<5,9>,P<6,17>,P<10,13>,P<11,12>,P<14,19>,P<15,18>>,
+                          L<P<0,2>,P<3,6>,P<4,5>,P<7,10>,P<8,11>,P<9,15>,P<12,16>,P<13,18>,P<14,17>,P<19,20>>,
+                          L<P<0,1>,P<2,3>,P<5,9>,P<6,12>,P<7,8>,P<11,14>,P<13,15>,P<16,19>,P<17,18>>,
+                          L<P<1,2>,P<3,9>,P<6,13>,P<10,11>,P<12,15>,P<16,17>,P<18,19>>,
+                          L<P<1,4>,P<2,5>,P<3,7>,P<6,10>,P<8,9>,P<11,12>,P<13,14>,P<17,18>>,
+                          L<P<2,4>,P<5,6>,P<7,8>,P<9,11>,P<10,13>,P<12,15>,P<14,16>>,
+                          L<P<3,4>,P<5,7>,P<6,8>,P<9,10>,P<11,13>,P<12,14>,P<15,16>>,
+                          L<P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<14,15>,P<16,17>>>,
+                /* 22 */ N<L<P<0,1>,P<2,3>,P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<14,15>,P<16,17>,P<18,19>,P<20,21>>,
+                          L<P<0,12>,P<1,13>,P<2,6>,P<3,7>,P<4,10>,P<8,20>,P<9,21>,P<11,17>,P<14,18>,P<15,19>>,
+                          L<P<0,2>,P<1,6>,P<3,12>,P<4,16>,P<5,17>,P<7,13>,P<8,14>,P<9,18>,P<15,20>,P<19,21>>,
+                          L<P<0,8>,P<1,15>,P<2,14>,P<3,9>,P<5,11>,P<6,20>,P<7,19>,P<10,16>,P<12,18>,P<13,21>>,
+                          L<P<0,4>,P<1,10>,P<3,8>,P<5,9>,P<7,14>,P<11,20>,P<12,16>,P<13,18>,P<17,21>>,
+                          L<P<1,3>,P<2,5>,P<4,8>,P<6,9>,P<7,10>,P<11,14>,P<12,15>,P<13,17>,P<16,19>,P<18,20>>,
+                          L<P<2,4>,P<3,12>,P<5,8>,P<6,11>,P<9,18>,P<10,15>,P<13,16>,P<17,19>>,
+                          L<P<1,2>,P<3,4>,P<5,7>,P<6,12>,P<8,11>,P<9,15>,P<10,13>,P<14,16>,P<17,18>,P<19,20>>,
+                          L<P<2,3>,P<4,5>,P<7,12>,P<8,10>,P<9,14>,P<11,13>,P<16,17>,P<18,19>>,
+                          L<P<4,6>,P<5,8>,P<9,11>,P<10,12>,P<13,16>,P<15,17>>,
+                          L<P<3,4>,P<6,7>,P<9,10>,P<11,12>,P<14,15>,P<17,18>>,
+                          L<P<5,6>,P<7,8>,P<10,11>,P<13,14>,P<15,16>>,
+                          L<P<6,7>,P<8,9>,P<12,13>,P<14,15>>>,
+                /* 23 */ N<L<P<0,20>,P<1,12>,P<2,16>,P<4,6>,P<5,10>,P<7,21>,P<8,14>,P<9,15>,P<11,22>,P<13,18>,P<17,19>>,
+                          L<P<0,3>,P<1,11>,P<2,7>,P<4,17>,P<5,13>,P<6,19>,P<8,9>,P<10,18>,P<12,22>,P<14,15>,P<16,21>>,
+                          L<P<0,1>,P<2,4>,P<3,12>,P<5,8>,P<6,9>,P<7,10>,P<11,20>,P<13,16>,P<14,17>,P<15,18>,P<19,21>>,
+                          L<P<2,5>,P<4,8>,P<6,11>,P<7,14>,P<9,16>,P<12,17>,P<15,19>,P<18,21>>,
+                          L<P<1,8>,P<3,14>,P<4,7>,P<9,20>,P<10,12>,P<11,13>,P<15,22>,P<16,19>>,
+                          L<P<0,7>,P<1,5>,P<3,4>,P<6,11>,P<8,15>,P<9,14>,P<10,13>,P<12,17>,P<18,22>,P<19,20>>,
+                          L<P<0,2>,P<1,6>,P<4,7>,P<5,9>,P<8,10>,P<13,15>,P<14,18>,P<16,19>,P<17,22>,P<20,21>>,
+                          L<P<2,3>,P<4,5>,P<6,8>,P<7,9>,P<10,11>,P<12,13>,P<14,16>,P<15,17>,P<18,19>,P<21,22>>,
+                          L<P<1,2>,P<3,6>,P<4,10>,P<7,8>,P<9,11>,P<12,14>,P<13,19>,P<15,16>,P<17,20>>,
+                          L<P<2,3>,P<5,10>,P<6,7>,P<8,9>,P<13,18>,P<14,15>,P<16,17>,P<20,21>>,
+                          L<P<3,4>,P<5,7>,P<10,12>,P<11,13>,P<16,18>,P<19,20>>,
+                          L<P<4,6>,P<8,10>,P<9,12>,P<11,14>,P<13,15>,P<17,19>>,
+                          L<P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>>>,
+                /* 24 */ N<L<P<0,20>,P<1,12>,P<2,16>,P<3,23>,P<4,6>,P<5,10>,P<7,21>,P<8,14>,P<9,15>,P<11,22>,P<13,18>,P<17,19>>,
+                          L<P<0,3>,P<1,11>,P<2,7>,P<4,17>,P<5,13>,P<6,19>,P<8,9>,P<10,18>,P<12,22>,P<14,15>,P<16,21>,P<20,23>>,
+                          L<P<0,1>,P<2,4>,P<3,12>,P<5,8>,P<6,9>,P<7,10>,P<11,20>,P<13,16>,P<14,17>,P<15,18>,P<19,21>,P<22,23>>,
+                          L<P<2,5>,P<4,8>,P<6,11>,P<7,14>,P<9,16>,P<12,17>,P<15,19>,P<18,21>>,
+                          L<P<1,8>,P<3,14>,P<4,7>,P<9,20>,P<10,12>,P<11,13>,P<15,22>,P<16,19>>,
+                          L<P<0,7>,P<1,5>,P<3,4>,P<6,11>,P<8,15>,P<9,14>,P<10,13>,P<12,17>,P<16,23>,P<18,22>,P<19,20>>,
+                          L<P<0,2>,P<1,6>,P<4,7>,P<5,9>,P<8,10>,P<13,15>,P<14,18>,P<16,19>,P<17,22>,P<21,23>>,
+                          L<P<2,3>,P<4,5>,P<6,8>,P<7,9>,P<10,11>,P<12,13>,P<14,16>,P<15,17>,P<18,19>,P<20,21>>,
+                          L<P<1,2>,P<3,6>,P<4,10>,P<7,8>,P<9,11>,P<12,14>,P<13,19>,P<15,16>,P<17,20>,P<21,22>>,
+                          L<P<2,3>,P<5,10>,P<6,7>,P<8,9>,P<13,18>,P<14,15>,P<16,17>,P<20,21>>,
+                          L<P<3,4>,P<5,7>,P<10,12>,P<11,13>,P<16,18>,P<19,20>>,
+                          L<P<4,6>,P<8,10>,P<9,12>,P<11,14>,P<13,15>,P<17,19>>,
+                          L<P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>>>,
+                /* 25 */ N<L<P<0,2>,P<1,8>,P<3,18>,P<4,17>,P<5,20>,P<6,19>,P<7,9>,P<10,11>,P<12,13>,P<14,16>,P<15,22>,P<21,23>>,
+                          L<P<0,3>,P<1,15>,P<2,18>,P<4,12>,P<5,21>,P<6,10>,P<7,14>,P<8,22>,P<9,16>,P<11,19>,P<13,17>,P<20,23>>,
+                          L<P<0,4>,P<1,7>,P<2,13>,P<3,12>,P<5,6>,P<8,14>,P<9,15>,P<10,21>,P<11,20>,P<16,22>,P<17,18>,P<19,23>>,
+                          L<P<0,5>,P<2,11>,P<3,6>,P<4,10>,P<7,16>,P<8,9>,P<12,21>,P<13,19>,P<14,15>,P<17,20>,P<18,23>>,
+                          L<P<2,7>,P<6,9>,P<8,11>,P<14,24>,P<18,21>>,
+                          L<P<3,8>,P<7,10>,P<11,12>,P<13,14>,P<15,21>,P<18,20>,P<22,24>>,
+                          L<P<4,13>,P<10,16>,P<11,15>,P<18,24>,P<19,22>>,
+                          L<P<1,4>,P<8,11>,P<9,19>,P<13,17>,P<14,18>,P<16,20>,P<23,24>>,
+                          L<P<0,1>,P<4,5>,P<6,13>,P<9,14>,P<10,17>,P<12,16>,P<18,19>,P<20,21>,P<22,23>>,
+                          L<P<2,6>,P<3,4>,P<5,13>,P<7,9>,P<12,18>,P<15,17>,P<16,19>,P<20,22>,P<21,23>>,
+                          L<P<1,2>,P<5,8>,P<6,7>,P<9,10>,P<11,13>,P<14,15>,P<17,20>,P<21,22>>,
+                          L<P<1,3>,P<2,4>,P<5,6>,P<7,11>,P<8,9>,P<10,13>,P<12,14>,P<15,16>,P<17,18>,P<19,20>>,
+                          L<P<2,3>,P<4,8>,P<6,7>,P<9,12>,P<10,11>,P<13,14>,P<15,17>,P<16,18>,P<20,21>>,
+                          L<P<3,5>,P<4,6>,P<7,8>,P<9,10>,P<11,12>,P<13,15>,P<14,17>,P<16,19>>,
+                          L<P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<14,15>,P<16,17>,P<18,19>>>,
+                /* 26 */ N<L<P<0,25>,P<1,3>,P<2,9>,P<4,19>,P<5,18>,P<6,21>,P<7,20>,P<8,10>,P<11,12>,P<13,14>,P<15,17>,P<16,23>,P<22,24>>,
+                          L<P<1,4>,P<2,16>,P<3,19>,P<5,13>,P<6,22>,P<7,11>,P<8,15>,P<9,23>,P<10,17>,P<12,20>,P<14,18>,P<21,24>>,
+                          L<P<1,5>,P<2,8>,P<3,14>,P<4,13>,P<6,7>,P<9,15>,P<10,16>,P<11,22>,P<12,21>,P<17,23>,P<18,19>,P<20,24>>,
+                          L<P<0,10>,P<1,6>,P<3,7>,P<4,11>,P<5,12>,P<13,20>,P<14,21>,P<15,25>,P<18,22>,P<19,24>>,
+                          L<P<0,4>,P<8,10>,P<12,13>,P<15,17>,P<21,25>>,
+                          L<P<0,2>,P<4,8>,P<10,12>,P<13,15>,P<17,21>,P<23,25>>,
+                          L<P<0,1>,P<2,3>,P<4,5>,P<8,14>,P<9,13>,P<11,17>,P<12,16>,P<20,21>,P<22,23>,P<24,25>>,
+                          L<P<1,4>,P<3,10>,P<6,9>,P<7,13>,P<8,11>,P<12,18>,P<14,17>,P<15,22>,P<16,19>,P<21,24>>,
+                          L<P<2,6>,P<3,8>,P<5,7>,P<9,12>,P<13,16>,P<17,22>,P<18,20>,P<19,23>>,
+                          L<P<1,2>,P<4,6>,P<5,9>,P<7,10>,P<11,12>,P<13,14>,P<15,18>,P<16,20>,P<19,21>,P<23,24>>,
+                          L<P<2,4>,P<3,5>,P<7,13>,P<8,9>,P<10,14>,P<11,15>,P<12,18>,P<16,17>,P<20,22>,P<21,23>>,
+                          L<P<3,4>,P<6,9>,P<7,11>,P<10,12>,P<13,15>,P<14,18>,P<16,19>,P<21,22>>,
+                          L<P<5,7>,P<6,8>,P<9,13>,P<10,11>,P<12,16>,P<14,15>,P<17,19>,P<18,20>>,
+                          L<P<5,6>,P<7,8>,P<9,10>,P<11,13>,P<12,14>,P<15,16>,P<17,18>,P<19,20>>,
+                          L<P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<14,15>,P<16,17>,P<18,19>,P<20,21>>>,
+                /* 27 */ N<L<P<0,9>,P<1,6>,P<2,4>,P<3,7>,P<5,8>,P<11,24>,P<12,23>,P<13,26>,P<14,25>,P<15,19>,P<16,17>,P<18,22>,P<20,21>>,
+                          L<P<0,1>,P<3,5>,P<4,10>,P<6,9>,P<7,8>,P<11,16>,P<12,18>,P<13,20>,P<14,15>,P<17,24>,P<19,25>,P<21,26>,P<22,23>>,
+                          L<P<1,3>,P<2,5>,P<4,7>,P<8,10>,P<11,12>,P<13,14>,P<15,16>,P<17,19>,P<18,20>,P<21,22>,P<23,24>,P<25,26>>,
+                          L<P<0,4>,P<1,2>,P<3,7>,P<5,9>,P<6,8>,P<11,13>,P<12,14>,P<15,21>,P<16,22>,P<17,18>,P<19,20>,P<23,25>,P<24,26>>,
+                          L<P<0,1>,P<2,6>,P<4,5>,P<7,8>,P<9,10>,P<12,13>,P<14,23>,P<15,17>,P<16,18>,P<19,21>,P<20,22>,P<24,25>>,
+                          L<P<0,11>,P<2,4>,P<3,6>,P<5,7>,P<8,9>,P<12,15>,P<13,17>,P<16,19>,P<18,21>,P<20,24>,P<22,25>>,
+                          L<P<1,2>,P<3,4>,P<5,6>,P<7,8>,P<13,15>,P<14,17>,P<20,23>,P<22,24>>,
+                          L<P<1,12>,P<2,3>,P<4,5>,P<6,7>,P<14,16>,P<17,19>,P<18,20>,P<21,23>>,
+                          L<P<2,13>,P<14,15>,P<16,17>,P<18,19>,P<20,21>,P<22,23>>,
+                          L<P<3,14>,P<4,15>,P<5,16>,P<10,21>,P<17,18>,P<19,20>>,
+                          L<P<6,17>,P<7,18>,P<8,19>,P<9,20>,P<10,13>,P<14,22>,P<15,23>,P<16,24>>,
+                          L<P<6,10>,P<7,14>,P<8,11>,P<9,12>,P<17,25>,P<18,26>,P<19,23>,P<20,24>>,
+                          L<P<4,8>,P<5,9>,P<11,15>,P<12,16>,P<13,17>,P<18,22>,P<21,25>,P<24,26>>,
+                          L<P<2,4>,P<3,5>,P<6,8>,P<7,9>,P<10,11>,P<12,14>,P<13,15>,P<16,18>,P<17,19>,P<20,22>,P<21,23>,P<25,26>>,
+                          L<P<1,2>,P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>,P<19,20>,P<21,22>,P<23,24>>>,
+                /* 28 */ N<L<P<0,9>,P<1,20>,P<2,21>,P<3,22>,P<4,19>,P<5,24>,P<6,25>,P<7,26>,P<8,23>,P<10,15>,P<11,13>,P<12,17>,P<14,16>,P<18,27>>,
+                          L<P<0,18>,P<1,7>,P<2,6>,P<3,5>,P<4,8>,P<9,27>,P<10,12>,P<11,14>,P<13,16>,P<15,17>,P<19,23>,P<20,26>,P<21,25>,P<22,24>>,
+                          L<P<1,2>,P<3,4>,P<5,19>,P<6,20>,P<7,21>,P<8,22>,P<9,18>,P<10,11>,P<12,14>,P<13,15>,P<16,17>,P<23,24>,P<25,26>>,
+                          L<P<0,3>,P<1,10>,P<5,8>,P<6,7>,P<11,13>,P<14,16>,P<17,26>,P<19,22>,P<20,21>,P<24,27>>,
+                          L<P<0,1>,P<2,7>,P<3,10>,P<4,8>,P<12,13>,P<14,15>,P<17,24>,P<19,23>,P<20,25>,P<26,27>>,
+                          L<P<1,3>,P<2,6>,P<4,5>,P<7,19>,P<8,20>,P<11,12>,P<13,14>,P<15,16>,P<21,25>,P<22,23>,P<24,26>>,
+                          L<P<2,4>,P<5,12>,P<7,8>,P<9,11>,P<10,14>,P<13,17>,P<15,22>,P<16,18>,P<19,20>,P<23,25>>,
+                          L<P<2,9>,P<4,11>,P<5,6>,P<7,13>,P<8,10>,P<14,20>,P<16,23>,P<17,19>,P<18,25>,P<21,22>>,
+                          L<P<1,2>,P<3,16>,P<4,9>,P<6,12>,P<10,14>,P<11,24>,P<13,17>,P<15,21>,P<18,23>,P<25,26>>,
+                          L<P<2,8>,P<3,5>,P<4,7>,P<6,16>,P<9,15>,P<11,21>,P<12,18>,P<19,25>,P<20,23>,P<22,24>>,
+                          L<P<2,3>,P<5,8>,P<7,9>,P<11,15>,P<12,16>,P<18,20>,P<19,22>,P<24,25>>,
+                          L<P<6,8>,P<10,12>,P<11,13>,P<14,16>,P<15,17>,P<19,21>>,
+                          L<P<5,6>,P<8,10>,P<9,11>,P<12,13>,P<14,15>,P<16,18>,P<17,19>,P<21,22>>,
+                          L<P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,14>,P<13,15>,P<16,17>,P<18,19>,P<20,21>,P<22,23>>,
+                          L<P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>,P<19,20>,P<21,22>,P<23,24>>>,
+                /* 29 */ N<L<P<0,12>,P<1,10>,P<2,9>,P<3,7>,P<5,11>,P<6,8>,P<13,26>,P<14,25>,P<15,28>,P<16,27>,P<17,21>,P<18,19>,P<20,24>,P<22,23>>,
+                          L<P<1,6>,P<2,3>,P<4,11>,P<7,9>,P<8,10>,P<13,18>,P<14,20>,P<15,22>,P<16,17>,P<19,26>,P<21,27>,P<23,28>,P<24,25>>,
+                          L<P<0,4>,P<1,2>,P<3,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>,P<19,21>,P<20,22>,P<23,24>,P<25,26>,P<27,28>>,
+                          L<P<4,6>,P<5,9>,P<8,11>,P<10,12>,P<13,15>,P<14,16>,P<17,23>,P<18,24>,P<19,20>,P<21,22>,P<25,27>,P<26,28>>,
+                          L<P<0,5>,P<3,8>,P<4,7>,P<6,11>,P<9,10>,P<14,15>,P<16,25>,P<17,19>,P<18,20>,P<21,23>,P<22,24>,P<26,27>>,
+                          L<P<0,1>,P<2,5>,P<6,9>,P<7,8>,P<10,11>,P<14,17>,P<15,19>,P<18,21>,P<20,23>,P<22,26>,P<24,27>>,
+                          L<P<0,13>,P<1,3>,P<2,4>,P<5,6>,P<9,10>,P<15,17>,P<16,19>,P<22,25>,P<24,26>>,
+                          L<P<1,2>,P<3,4>,P<5,7>,P<6,8>,P<16,18>,P<19,21>,P<20,22>,P<23,25>>,
+                          L<P<1,14>,P<2,3>,P<4,5>,P<6,7>,P<8,9>,P<16,17>,P<18,19>,P<20,21>,P<22,23>,P<24,25>>,
+                          L<P<2,15>,P<3,4>,P<5,6>,P<10,23>,P<11,24>,P<12,25>,P<19,20>,P<21,22>>,
+                          L<P<3,16>,P<4,17>,P<5,18>,P<6,19>,P<7,20>,P<8,21>,P<9,22>,P<10,15>>,
+                          L<P<6,10>,P<8,13>,P<9,14>,P<11,16>,P<12,17>,P<18,26>,P<19,27>,P<20,28>>,
+                          L<P<4,8>,P<5,9>,P<7,11>,P<12,13>,P<14,18>,P<15,19>,P<16,20>,P<17,21>,P<22,26>,P<23,27>,P<24,28>>,
+                          L<P<2,4>,P<3,5>,P<6,8>,P<7,9>,P<10,12>,P<11,14>,P<13,15>,P<16,18>,P<17,19>,P<20,22>,P<21,23>,P<24,26>,P<25,27>>,
+                          L<P<1,2>,P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>,P<19,20>,P<21,22>,P<23,24>,P<25,26>,P<27,28>>>,
+                /* 30 */ N<L<P<1,2>,P<3,10>,P<4,14>,P<5,8>,P<6,13>,P<7,12>,P<9,11>,P<16,17>,P<18,25>,P<19,29>,P<20,23>,P<21,28>,P<22,27>,P<24,26>>,
+                          L<P<0,14>,P<1,5>,P<2,8>,P<3,7>,P<6,9>,P<10,12>,P<11,13>,P<15,29>,P<16,20>,P<17,23>,P<18,22>,P<21,24>,P<25,27>,P<26,28>>,
+                          L<P<0,7>,P<1,6>,P<2,9>,P<4,10>,P<5,11>,P<8,13>,P<12,14>,P<15,22>,P<16,21>,P<17,24>,P<19,25>,P<20,26>,P<23,28>,P<27,29>>,
+                          L<P<0,6>,P<2,4>,P<3,5>,P<7,11>,P<8,10>,P<9,12>,P<13,14>,P<15,21>,P<17,19>,P<18,20>,P<22,26>,P<23,25>,P<24,27>,P<28,29>>,
+                          L<P<0,3>,P<1,2>,P<4,7>,P<5,9>,P<6,8>,P<10,11>,P<12,13>,P<14,29>,P<15,18>,P<16,17>,P<19,22>,P<20,24>,P<21,23>,P<25,26>,P<27,28>>,
+                          L<P<0,1>,P<2,3>,P<4,6>,P<7,9>,P<10,12>,P<11,13>,P<15,16>,P<17,18>,P<19,21>,P<22,24>,P<25,27>,P<26,28>>,
+                          L<P<0,15>,P<1,2>,P<3,5>,P<8,10>,P<11,12>,P<13,28>,P<16,17>,P<18,20>,P<23,25>,P<26,27>>,
+                          L<P<1,16>,P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<12,27>,P<18,19>,P<20,21>,P<22,23>,P<24,25>>,
+                          L<P<2,3>,P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<17,18>,P<19,20>,P<21,22>,P<23,24>,P<25,26>>,
+                          L<P<2,17>,P<3,18>,P<4,19>,P<5,6>,P<7,8>,P<9,24>,P<10,25>,P<11,26>,P<20,21>,P<22,23>>,
+                          L<P<5,20>,P<6,21>,P<7,22>,P<8,23>,P<9,16>,P<10,17>,P<11,18>,P<12,19>>,
+                          L<P<5,9>,P<6,10>,P<7,11>,P<8,15>,P<13,20>,P<14,21>,P<18,22>,P<19,23>>,
+                          L<P<3,5>,P<4,8>,P<7,9>,P<12,15>,P<13,16>,P<14,17>,P<20,24>,P<21,25>>,
+                          L<P<2,4>,P<6,8>,P<10,12>,P<11,13>,P<14,15>,P<16,18>,P<17,19>,P<20,22>,P<21,23>,P<24,26>,P<25,27>>,
+                          L<P<1,2>,P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>,P<19,20>,P<21,22>,P<23,24>,P<25,26>,P<27,28>>>,
+                /* 31 */ N<L<P<0,1>,P<2,3>,P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<14,15>,P<16,17>,P<18,19>,P<20,21>,P<22,23>,P<24,25>,P<26,27>,P<28,29>>,
+                          L<P<0,2>,P<1,3>,P<4,6>,P<5,7>,P<8,10>,P<9,11>,P<12,14>,P<13,15>,P<16,18>,P<17,19>,P<20,22>,P<21,23>,P<24,26>,P<25,27>,P<28,30>>,
+                          L<P<0,4>,P<1,5>,P<2,6>,P<3,7>,P<8,12>,P<9,13>,P<10,14>,P<11,15>,P<16,20>,P<17,21>,P<18,22>,P<19,23>,P<24,28>,P<25,29>,P<26,30>>,
+                          L<P<0,8>,P<1,9>,P<2,10>,P<3,11>,P<4,12>,P<5,13>,P<6,14>,P<7,15>,P<16,24>,P<17,25>,P<18,26>,P<19,27>,P<20,28>,P<21,29>,P<22,30>>,
+                          L<P<0,16>,P<1,8>,P<2,4>,P<3,12>,P<5,10>,P<6,9>,P<7,14>,P<11,13>,P<17,24>,P<18,20>,P<19,28>,P<21,26>,P<22,25>,P<23,30>,P<27,29>>,
+                          L<P<1,2>,P<3,5>,P<4,8>,P<6,22>,P<7,11>,P<9,25>,P<10,12>,P<13,14>,P<17,18>,P<19,21>,P<20,24>,P<23,27>,P<26,28>,P<29,30>>,
+                          L<P<1,17>,P<2,18>,P<3,19>,P<4,20>,P<5,10>,P<7,23>,P<8,24>,P<11,27>,P<12,28>,P<13,29>,P<14,30>,P<21,26>>,
+                          L<P<3,17>,P<4,16>,P<5,21>,P<6,18>,P<7,9>,P<8,20>,P<10,26>,P<11,23>,P<13,25>,P<14,28>,P<15,27>,P<22,24>>,
+                          L<P<1,4>,P<3,8>,P<5,16>,P<7,17>,P<9,21>,P<10,22>,P<11,19>,P<12,20>,P<14,24>,P<15,26>,P<23,28>,P<27,30>>,
+                          L<P<2,5>,P<7,8>,P<9,18>,P<11,17>,P<12,16>,P<13,22>,P<14,20>,P<15,19>,P<23,24>,P<26,29>>,
+                          L<P<2,4>,P<6,12>,P<9,16>,P<10,11>,P<13,17>,P<14,18>,P<15,22>,P<19,25>,P<20,21>,P<27,29>>,
+                          L<P<5,6>,P<8,12>,P<9,10>,P<11,13>,P<14,16>,P<15,17>,P<18,20>,P<19,23>,P<21,22>,P<25,26>>,
+                          L<P<3,5>,P<6,7>,P<8,9>,P<10,12>,P<11,14>,P<13,16>,P<15,18>,P<17,20>,P<19,21>,P<22,23>,P<24,25>,P<26,28>>,
+                          L<P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>,P<19,20>,P<21,22>,P<23,24>,P<25,26>,P<27,28>>>,
+                /* 32 */ N<L<P<0,1>,P<2,3>,P<4,5>,P<6,7>,P<8,9>,P<10,11>,P<12,13>,P<14,15>,P<16,17>,P<18,19>,P<20,21>,P<22,23>,P<24,25>,P<26,27>,P<28,29>,P<30,31>>,
+                          L<P<0,2>,P<1,3>,P<4,6>,P<5,7>,P<8,10>,P<9,11>,P<12,14>,P<13,15>,P<16,18>,P<17,19>,P<20,22>,P<21,23>,P<24,26>,P<25,27>,P<28,30>,P<29,31>>,
+                          L<P<0,4>,P<1,5>,P<2,6>,P<3,7>,P<8,12>,P<9,13>,P<10,14>,P<11,15>,P<16,20>,P<17,21>,P<18,22>,P<19,23>,P<24,28>,P<25,29>,P<26,30>,P<27,31>>,
+                          L<P<0,8>,P<1,9>,P<2,10>,P<3,11>,P<4,12>,P<5,13>,P<6,14>,P<7,15>,P<16,24>,P<17,25>,P<18,26>,P<19,27>,P<20,28>,P<21,29>,P<22,30>,P<23,31>>,
+                          L<P<0,16>,P<1,8>,P<2,4>,P<3,12>,P<5,10>,P<6,9>,P<7,14>,P<11,13>,P<15,31>,P<17,24>,P<18,20>,P<19,28>,P<21,26>,P<22,25>,P<23,30>,P<27,29>>,
+                          L<P<1,2>,P<3,5>,P<4,8>,P<6,22>,P<7,11>,P<9,25>,P<10,12>,P<13,14>,P<17,18>,P<19,21>,P<20,24>,P<23,27>,P<26,28>,P<29,30>>,
+                          L<P<1,17>,P<2,18>,P<3,19>,P<4,20>,P<5,10>,P<7,23>,P<8,24>,P<11,27>,P<12,28>,P<13,29>,P<14,30>,P<21,26>>,
+                          L<P<3,17>,P<4,16>,P<5,21>,P<6,18>,P<7,9>,P<8,20>,P<10,26>,P<11,23>,P<13,25>,P<14,28>,P<15,27>,P<22,24>>,
+                          L<P<1,4>,P<3,8>,P<5,16>,P<7,17>,P<9,21>,P<10,22>,P<11,19>,P<12,20>,P<14,24>,P<15,26>,P<23,28>,P<27,30>>,
+                          L<P<2,5>,P<7,8>,P<9,18>,P<11,17>,P<12,16>,P<13,22>,P<14,20>,P<15,19>,P<23,24>,P<26,29>>,
+                          L<P<2,4>,P<6,12>,P<9,16>,P<10,11>,P<13,17>,P<14,18>,P<15,22>,P<19,25>,P<20,21>,P<27,29>>,
+                          L<P<5,6>,P<8,12>,P<9,10>,P<11,13>,P<14,16>,P<15,17>,P<18,20>,P<19,23>,P<21,22>,P<25,26>>,
+                          L<P<3,5>,P<6,7>,P<8,9>,P<10,12>,P<11,14>,P<13,16>,P<15,18>,P<17,20>,P<19,21>,P<22,23>,P<24,25>,P<26,28>>,
+                          L<P<3,4>,P<5,6>,P<7,8>,P<9,10>,P<11,12>,P<13,14>,P<15,16>,P<17,18>,P<19,20>,P<21,22>,P<23,24>,P<25,26>,P<27,28>>>
+    >;
+    // clang-format on
+  }  // namespace sorter_hunter
+
+  // Generates sorting networks which are either proven to be
+  // size optimal (and are thus comprised of minimum number of compare
+  // and swap operations) or are found by the "SorterHunter"
+  // algorithm (https://github.com/bertdobbelaere/SorterHunter)
+  template<std::size_t N>
+  struct sorting_network<N, type::size_optimized_sort,
+                         std::enable_if_t<(N < std::tuple_size<sorter_hunter::Networks>::value)>>
+  {
+    using type = std::tuple_element_t<N - 1, sorter_hunter::Networks>;
+  };
+}  // namespace quxflux::sorting_net
